# src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Commune API",
    "version": "1.0.0",
    "description": "The Commune API lets AI agents and developers send and receive email, manage inboxes and domains, and process structured data from email conversations. All endpoints require a Bearer API key.\n\nBase URL: `https://api.commune.email`\n\nAuthentication: Pass your API key as `Authorization: Bearer comm_...`",
    "contact": {
      "name": "Commune Support",
      "url": "https://commune.email",
      "email": "support@commune.email"
    },
    "license": {
      "name": "Business Source License 1.1",
      "url": "https://mariadb.com/bsl11/"
    }
  },
  "servers": [
    {
      "url": "https://api.commune.email",
      "description": "Production"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "comm_...",
        "description": "API key issued from the Commune dashboard. Prefix `comm_`."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message"
          }
        }
      },
...
      "Attachment": {
        "type": "object",
        "required": ["attachment_id", "filename", "mime_type", "size"],
        "properties": {
          "attachment_id": { "type": "string" },
          "message_id": { "type": "string" },
          "filename": { "type": "string" },
          "mime_type": { "type": "string" },
          "size": { "type": "integer", "description": "File size in bytes" },
          "storage_type": { "type": "string", "enum": ["database", "cloudinary"] },
          "cloudinary_url": { "type": "string", "nullable": true }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid API key",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "Unauthorized" }
          }
        }
      },
...
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "BadRequest": {
        "description": "Invalid request body or parameters",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
      "InternalError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      }
    }
  },
  "paths": {
...
    "/v1/attachments/upload": {
      "post": {
        "tags": ["Attachments"],
        "operationId": "uploadAttachment",
        "summary": "Upload an attachment",
        "description": "Upload a file to use as an attachment when sending emails. Returns an `attachment_id` to pass in the `attachments` array of `POST /v1/messages/send`.\n\nMaximum file size is limited by your plan's storage quota.\n\nRequires `attachments:write` permission.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["content", "filename", "mime_type"],
                "properties": {
                  "content": {
                    "type": "string",
                    "description": "Base64-encoded file content"
                  },
                  "filename": {
                    "type": "string",
                    "description": "Original filename, e.g. `invoice.pdf`"
                  },
                  "mime_type": {
                    "type": "string",
                    "description": "MIME type, e.g. `application/pdf`"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Attachment uploaded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "attachment_id": { "type": "string" },
                        "filename": { "type": "string" },
                        "mime_type": { "type": "string" },
                        "size": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "413": {
            "description": "Attachment storage quota exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string" },
                    "used_mb": { "type": "number" },
                    "limit_mb": { "type": "number" },
                    "upgrade_url": { "type": "string" }
                  }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/v1/attachments/{attachmentId}": {
      "get": {
        "tags": ["Attachments"],
        "operationId": "getAttachment",
        "summary": "Get attachment metadata",
        "description": "Returns metadata for an attachment without its content. Use `GET /v1/attachments/{attachmentId}/url` to get a download link.\n\nRequires `attachments:read` permission.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "attachmentId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Attachment metadata",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/Attachment" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/v1/attachments/{attachmentId}/url": {
      "get": {
        "tags": ["Attachments"],
        "operationId": "getAttachmentUrl",
        "summary": "Get attachment download URL",
        "description": "Returns a temporary signed URL to download the attachment. For Cloudinary-stored files the URL expires after `expires_in` seconds. For database-stored files, a `data:` URL is returned.\n\nRequires `attachments:read` permission.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "attachmentId", "in": "path", "required": true, "schema": { "type": "string" } },
          {
            "name": "expires_in",
            "in": "query",
            "description": "URL expiration in seconds (default 3600)",
            "schema": { "type": "integer", "default": 3600 }
          }
        ],
        "responses": {
          "200": {
            "description": "Download URL",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "url": { "type": "string" },
                        "expires_in": { "type": "integer" },
                        "filename": { "type": "string" },
                        "mime_type": { "type": "string" },
                        "size": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
    "/v1/credits": {
      "get": {
        "tags": ["Credits"],
        "operationId": "getCredits",
        "summary": "Get credit balance",
        "description": "Returns the current SMS/voice credit balance for your organization.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Credit balance",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "balance": { "type": "number", "description": "Available credits" },
                        "lifetime_credits": { "type": "number" },
                        "lifetime_used": { "type": "number" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/v1/credits/bundles": {
      "get": {
        "tags": ["Credits"],
        "operationId": "listCreditBundles",
        "summary": "List credit bundles",
        "description": "Returns available credit purchase bundles with pricing.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Credit bundles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string", "enum": ["starter", "growth", "scale"] },
                          "credits": { "type": "number" },
                          "price": { "type": "number", "description": "Price in USD" },
                          "price_per_credit": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/v1/credits/checkout": {
      "post": {
        "tags": ["Credits"],
        "operationId": "createCreditsCheckout",
        "summary": "Create credit purchase checkout",
        "description": "Create a Stripe Checkout session to purchase a credit bundle. Returns a `checkout_url` to redirect the user to.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["bundle"],
                "properties": {
                  "bundle": {
                    "type": "string",
                    "enum": ["starter", "growth", "scale"]
                  },
                  "success_url": { "type": "string", "format": "uri", "description": "Redirect URL after successful payment" },
                  "cancel_url": { "type": "string", "format": "uri", "description": "Redirect URL if checkout is cancelled" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Checkout session",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "checkout_url": { "type": "string", "format": "uri" },
                        "bundle": { "type": "string" },
                        "credits": { "type": "number" },
                        "price": { "type": "number" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  },
...
}